Identifier: if you skip one, don't come back to it until the end
A bit incomplete b/c some things I don't know how to do.  But this is a
start about how I'd like it to work.  If you skip one, you don't get it
again (unless there are no more new ones, then it should go back).  This
is implemented.

Not yet implemented: if you manually go back and do a skipped one, it should
go to the next one in the table (rather than jumping to the bottom).

Also some monkey patching of the UI which should probably done elsewhere.