Refactor the upQ callbacks
First, talk to the server inside upQ itself.  Then emit either of two
signals, one for success and one for failure.  After much debugging,
I feel its bad to open dialogs in the thread, crashy-crashy and hard
to debug.  Eventually what works for me was refactoring this particular
call to SRMsg to not make popups: this is better anyway b/c otherwise
the user sees two dialogs with some repetition (which is confusing).