messenger: protect simultaneous calls to SRMsg with a mutex
Its not clear there is any harm in simultaneous calls.  But given that
I don't know and it might've lead to crash before, I think its safer to
prevent simultaneous access from multiple threads.  SRMsg calls are fast
so shouldn't be any blocking from this...  Of course, if you want some
blocking for debugging, its easy to sleep inside the mutex.  I left debug
code in place---but commented out---which does this.