Throw specific nomore exception
Callers can deal with however they deal with the superclass (BenignException).
So no change to callers is urgently needed (Identify and Totaller are not
touched by this change.  OTOH, in the commit, Marker is changed to catch
this particular exception and do nothing (code to popup dialog instead is
commented out.)