select no longer moved, and update view on "Done"
The background downloader no longer moves the row selection: we move
it when we're ready, so no need for special cases on cancel.  We do
need to update the view on "Done" however.  Finally, don't reset the
view unless we're actually going to move the selection (i.e., on Next)