hide the animating progress bars when no papers
We use a custom formatting for the text label.  Probably for now we
still need the dialog b/c its not clear that all Qt styles actually
show the labels.

A future improvement would be to write something into the titlebar
or into the "paper list" widget.  Then we can remove the error dialog.