Marker should display progress as XofY
It always did before (it is specified in the .ui file).  But after
8d5bbba740299e6bfe6b4b8e4bd3db0ffd9cbb79 we call "resetFormat()" which
annoyingly doesn't reset back to object creation style but rather to
the default (which on most themes seems to be percentage).  Ok, so
hardcode it here too?  Well that's not great either:
  * hardcoded in two places
  * not translated
So as a workaround, we grab whatever formatting we had at time of
 `Marker.__init__` and cache that in a hidden variable.