ann/marker: disentangle the viewWholePage feature
Now marker doesn't store part of annotators state.  We ask marker to
download some stuff for us.  It gives us back a list of files.  When
we're done with them we pass the list back to Marker (in case Marker
wants to cache them---for now it just deletes them same as before this
change).