Note "for this session"
Current behaviour: user can explicit disable/enable these warnings in
the choose options UI.  This has global effect across sessions.

The actual dialog that pops up has a "don't show this again" checkbox.
The current behaviour is that this is *for this session*.  So this
commit changes the checkbox text to match that behaviour.

We could connect it directly to the settings and then change it back to
"Don't ask me again".  But maybe we like it like this... to discuss...