scene: first attempt to draw images at the same height
This isn't right because it scales down images.  What I want to do is
set the image height to be say 2000 and then something (the painter?)
renders it at an appropriate resolution based on the current zoom.
What this does (I think) is rescale the image to 2000px height and then
blit that onto the scene.  But its a start: at least all the images have
the same apparent height.