Separate "sidebar on right" from left-hand mouse
In my opinion, sidebar on right is more of "right-to-left" language
thing rather than a handedness thing.  In the past, we disagreed on
this.  So this commit simply decouples the option.

The default is right-hand mouse and left-hand sidebar.  If you change to
left-hand mouse, the side by default stays on the left but the buttons
re-order to match the right-hand keyboard shortcuts.  All four
combinations of these two toggles work and make sense.