show busy cursor when waiting for page rearranger
It can download quite a bit of stuff so there is often a noticiable
pause before the window opens.  The UI feels more responsive if we show
a wait cursor.  But to show a wait cursor, we need to let the event loop
process.  This is bad because you can click around during this time.
Therefore, I disable the entire Annotator UI during the time that Page
Rearranger is open.

Maybe there are better patterns for responsive Qt UIs that don't lock
when you click a button (such as open the dialog instantly and the
dialog displays spinners while each image downloads).  Leave a TODO for
future improvements.