patch the leak by clearing the lists on dialog close
It seems necessary to explicitly clear the source and sink lists.  Not
sure why yet.  and this not the right place to do so but it seems to
work for me.  Related to #1322.

This seems to fix it perfectly or nearabouts in artificial testing
(calling the dialog in a tight loop from an artificial driver
functions).  But in practice I still get some growth on repeated ctrl-R
calls.