Remove TensorFlow from deps
Code is still present: someone in the future still needs to check which
of the scikitlearn or tensorflow gives better results for our digit
recognition problem on large-scale data.  In particular, the
scikit-learn implementation was only ever tested on MNIST data, which is
what is is trained on so that's not really a valid test.

Fixes Issue #1346.  Well, for weak definition of "Fixes".