Prefer centralized locations for config/comments
Use a new dep on `appdirs` to calculate these locations.  For now, we
try to load the files from local dir first.  This could be dangerous as
a local file would then overwrite your user-configs.  At least for
comments, a merge might be better.  But comments should be going online
Real Soon anyway...