Early pull of all pagedata, pgadjust avoids api call
It still downloads: this doesn't do the "aggressive downloader" idea yet
but it does mean that in the all-pg-to-all-question variety of
hwscanning, "Adjust Pages" becomes almost instantaneous and avoids API
calls at a busy time (background upload and download are both going if
user opens "Adjust Pages" early).

This implementation isn't great: we basically download the full page
metadata in the three places we can download stuff.  Its likely the
existing `src_img_data` could be refactored to simplify all this.