Fix race by removing (temporarily?) the singleshot.
<rant>We don't know how to make multithreaded Qt app.  How hard can it
be to add an object with some "..." state (say animating), then call an
API in a background thread, then replace the "..." animating thing with
the result of the API call?</rant>

Apparently hard b/c I just spend 3 hours bisecting to find this!

This Fixes #1391.  I think.  But it gives a worse user experience.