Comment: add comments where blurb forces text-to-png
Text (blurb) can (and current does) hide the text->to->PNG latency
behind a qtimer.  But this deferral means the object has the wrong size
when we do tweakPositions.  So until that is more reactive to size
changes, we need to keep this `textToPng()` call, even though Blurb is
going to do it itself in 5 ms or so.  Document all this for future
cleanups.