Don't store state in instance until return
It looks like RubricWrangler put some stuff in self.wranglerState, then
put it into widgets and what not and then just left it stale until it
was set on dialog accept.  I think its better to just pass a variable
around to set the initial state.  And create the instance variable only
on dialog accept (as that seems to be a common mechanism for copying
data out of dialogs).