Proper message back to client not just UHT
A missing comma and some confusion in the return types of the
giveUserToken function meant that the server was just replying with
"UHT" for "User Has Token" I suppose.  But other cases, there was a
human-readable error message, but it was garbled like "APIPlom API...".
I've documented what I think the original intention was: a 3-tuple.

We can then clean up the handling of all this.  But in many places there
is special handling of the UHT case.  I think much of it could be
deleted now that the exception from Messenger gives better info.