Don't call scripts in demo and tests
We may not be installed, just running in source tree and tests should
still work.  Instead, we can call the underlying functions (good) or
things like `python3 -m plom.scripts.scan` (not so good).  Latter is not
portable to Windows where the executable is "python" for example.
Better to not use subprocess at all, but this is a larger effort to make
Plom more usuable from within Python.