Take initial orientation data from the full metadata
The "full_pagedata" has the server's initial orientation data.  Currently
this is always zero but later it could be used to to metadata rotations
on upload.  Anyway, this change reduces the use of the legacy
"page_metadata", so in some places drop that variable and use the
img_src_data instead.

Fixes #1306.