Avoid star imports and fix an old bug
It seems like the (manager?) act of reverting has been broken since at
least 2020-06, likely earlier.  Did not try to reproduce but these
variables were not defined so server thread should have crashed (after
the atomic lock so it probably had the desired effect).  Probably we
should check if there is a bug in manager that meant it never checked
the return code?