Download spec and compare to file after pingserver
I hit a problem where I already had a server locally on port 41984.  So
I started a new server and it failed with IOError, but b/c that is
running under a Popen, we don't know that it failed.  We can just sleep
and then check `process_is_running` (which will return False
eventually).  But if you query to early it will return True (b/c server
is still starting.  This is why I originally added the ping_server
method.  HOWEVER, in my case I end up pinging the wrong server and
erroneously concluding new server is alive!

A partial fix: ask the server for its spec and compare with the local
spec: specifically the `publicCode` should match.  This is still not
perfect b/c I could try to start two servers in the same existing
populated basedir.  But it least it catches some class of dumb
accidents.