Move fakepdf feature into the actual function
Too much indirection for me to handle: the _make_pdf frontend to
make_pdf is bad enough without it actually doing something.  The fakepdf
"feature" seems to be broken somewhere along the way, so this fixes
that.  Also, I started using "*," to mark the list inputs as kwarg-only
but this could complicate the parallel call (with list of args).