Docker: slick script too slick for own good
On Ubuntu 20.04, there is "or" logic in the depdencies:
```
qt5-default
 |Depends: qtbase5-dev
  Depends: qtbase5-gles-dev
  Depends: qtchooser
```
So the awk etc magic was trying to install both of an xor dependency.
This gives scary errors like:
```
This may mean that you have
requested an impossible situation or if you are using the unstable
...
```

So just hardcode it!