Exception handler dance on run not on import
Random things happening on import is sloppy.  Move `sys.exceptionhook`
hackery into a function that we can then call around the time where
other ctrl-C handlers and other nonsense is occuring.  I gave this a
dumb name because its seems like dumb thing that Qt should be doing for
us.  And if it has a dumb name that will inhibit use right?  Right?!

Also combine and dedupe Manager and Client code for this stuff.
Fixes #1621, because Client now has the fancy dialog.