Do some fake split in the upload implementation for testing
I'm not crazy about this, but I don't see how to maintain the mock test
without this split.  Maybe the mock test is not worth keeping but we
don't have any more of them...  So I'll leave it for now, with some
underscore names to make it clear no one should be using the `_raw` one.