Download 1, reassemble 1, rather than download all
Removes the parallel Pool bit so its much slower, maybe 2.5 s per paper
over the internet for a real test.  But no fork / file-handle bomb on
reassembly, so that's nice.

Fixes #1692.

Why can't we "just" pool/thread the one-at-a-time reassembler?  We
could, but I think it means cloning the msgr: not a big deal in
principle but this code was also hard to understand and maintain.  So a
simpler version that can be done in parallel?  Sure, send a MR.