More polite error message
The phrasing was awkward here implying you should do something before
trying again.  Server does not have an opinion here (we don't care if we
try again immediately or have a coffee first and ponder the meaning of
life).  So might as well be polite.

This is user-facing: the exact error text is shown to the user in a
dialog.