Qt font.setpixelsize wants ints: so round the float input
This is suddenly an error for me on Fedora 35.  Not sure why.  I tried a
PyQt5 5.15.0 and 5.15.6, same behaviour.  I might be the upgrade
from Python 3.9 to 3.10, perhaps a stricter typing to C code interface?

Anyway, the function definitely expects `int` so the main change here is
to round the float rather than chop (which is presumably what was
happening automatically before).