Claim Task was conflating all errors as PlomTaken
Fixes #1267.  The two main things that can go wrong here are: someone
else has claimed that task or the task does not (or not yet) exist.

I changed a 204 to a 409: mainly because I wasted a lot of time trying
to pass text with the 204 before realizing it literally means "no
content".  Anyway, elsewhere we use 4xx for non-owner conditions so I
think that should be ok here too.