Check version on validate and connection
We have the API check for formal "it won't work" checks.  This is more
of a warning: you have an old client which is compatible in the sense of
API but may have bugs that have been fixed in recent versions.
Fixes #1833.

A future improvement of this might have the server recommend a minimum
version.