Troublesome to have module and function with same name
Maybe we shouldn't be doing that...  But I want the command line to
be `python3 -m plom.produce.rando_exam_scribbler` but also be able to
call a function from within Python as:
```
from plom.produce import rando_exam_scribbler
```
But it seems this is perilous.

Instead use a `make_exam_scribbles` function, which is the functional
form of the the `rando_exam_scribbler`.  make scribbles versus
scribbler.  Sure to confuse me later, but at least somewhat similar to
the randoMarker versus do_rando_marking situation.