ask forgivenness instead of premission
I think this code is much easier to express with this pattern: instead
of checking if its scanned and deciding what to do, just try to move the
unknown, and then deal with the "scanned" failure.

Sadly, didn't get to delete much code cuz it looks like checkTPage is
used for other things too.