Fix when local rotation undoes DB rotation back to zero
If initial DB orientation is nonzero but local rotation is zero then we
still have work to do.  For example, say DB rotation is 90, but local
rotation is 0 (i.e., user rotated -90 from their PoV).  In this case,
we were incorrectly displaying 90.  Easiest fix is just to always do the
rotation.

Although I guess that is twice as expensive: we could skip zero
rotations in the actual rotation code if that is a concern (its not).