Generalize unknownPages info API call
It uses to just give a list of the (server-side) filenames of the
unknown pages.  Instead return a list of dicts, where each dict is a row
of various data.

Also, move this API call to BaseManager so scanner can use it and
(temporarily?) remove the restriction on manager-only calls.