Clarify if user exists, don't update password
This is how the code behaves but the warning message was unclear, and
led to some people thinking they could update the initial_user_list.json
to update passwords.  It could be implemented that way, perhaps with a
"--force" but for now it would be just too mysterious if your passwords
reset everytime you restarted the server.