Split prediction running from LAP solving
Don't use a thread/subprocess for the LAP solver, just run it inline.
Hopefully that means it runs within 60 seconds, including the time to
load all the probabilities from disc: that needs to be confirmed on
large dataset.  LAPsolvers docs say 3 seconds for 5000x5000 so that
aspect is not likely to be bottleneck.