Manager: if user wants to load them all, let them
We had some nanny code that intercepted the three star case.  The API
support all of them.  I guess there is a concern about bandwidth (?) or
more likely server load but given this is a manager call, I say we leave
that up to the caller.

Surely its no worse than some of the other manager API calls.