Reset the zoom fixed button on image reset
This fixes some papercuts in Marker when you ctrl-click to unselect a
paper, and means the Identifier zoom link resets properly when
switching between papers.  Note we don't really want it to reset between
pages in the case of Identifier but previous to this commit we had to
worst of both worlds: if you zoomed in and went next, it reset the zoom
back to zoomed out but locked it that way.