Annotator: proof of concept for single source keys
Our key-board shortcuts are spread out over all sorts of places.  This
shows how to use the toml file as a "single source of truth".  This
means we need "`ctrl+n`" not "`ctrl-n`": could do that by
postprocessing...  or just change the single-source as we do here.

Also need to update the displayed keys in the F10 "hamburger" menu.
In this case, we use the first key in the ordered list.