Fix for rubric creation
I think there is a key versus index mistake here.  This is a one-liner
fix but I had to clean up a bit to find it so have left the clean-up as
well.  The functional change here is:
```
-if Rubric.get_or_none(rID):
+if Rubric.get_or_none(key=rID):
```
That is, the thing called an ID is actually a key and those seem to be
two fields in the DB (who knows why).

Fixes #2006. Fixes #2269.