Set placeholder before download thread: prevent race
We did this before but first need to be careful about where we look for
the suggested file name: we were overwriting the `.svg` placeholder with
image data!  That was Issue #2388.  I don't think we have an issue
number for the potential (and likely extremele rare) race, but this is
the proper fix now that Issue #2388 is done.