Fix IDputPredictions() and remove checks
This commit illustrates that predictions are still uniquely assigned to
one row in the database, even without any additional checks or restrictions.
Both MLGreedy and MLLAP are called. A new invocation to the ML predictor
and subsequent database update overrides any old values in the table -
new rows are yet created.

Fixes #2080