Add both MLLAP and MLGreedy predictions into DB
db_create.py: if an id prediction has a predictor that is not
yet featured in the DB, then create a new row. Else, replace the
old prediction with the corresponding predictor.

Deletion and replacement of predictions: if lap_solver or greedy
are re-run, they will not yield different certainties, so there is
no comparison to be made with previous predictions. It is therefore
assumed that if the id predictor is re-run, the user would like to
either add more predictions or replace the old ones.

Only MLLAP appears in the table shown in Manager, but queries to the
database show that all predictions are added.