Draft for displaying all predictions in Identifier
Work in progress for #2444

Main ideas/changes: all predictions should be passed
around in a single dict that can store multiple
predictions for a single paper number. Iterate
through the dict and display the predictions
in Identifier as needed.

In this commit, certainties and both predictors are
displayed in a simple way in the Identifier client;
more UI changes to come.