rubrics: store ints for "value" and "out_of"
Tried of strings instead of numbers, so trying to see how far I can get
forcing ints.  They are now ints in the database (maybe floats some
day?) and ints in most of the dict representations.  This requires some
conversion back-and-forth in the Qt table storage.  But no conversions
during calculations.

Fixes #2447.