CI: apt-get update before installing packages
We don't know for sure when it was last done, it depends on whether
DOCKER_USE_CACHE was used, what branch, etc etc.  So just do it.

See Issue #2468.  I suppose this Fixes #2468.  Although, there maybe
some other effects of caching.  We should still DOCKER_USE_CACHE=0
periodically, especially before releases and on the main branch.  As we
do with a scheduled CI run.