Qt's setFocus does not take a bool
This probably works b/c it can take some kind of enum, so we get
whatever "1" is in that enum for `True` and "0" for `False` or something
undesirable like that.

The `setFocus(False)` seems just plain wrong: I would say it was
actually SETTING the focus with wtf knows as the FocusReason.  Thus I'm not
sure the bit in pagescene is doing anything...  But anyway, I set it to
`X.clearFocus()` for now.