LAP: add certainty, move assemble_cost_matrix call
Fixes #2440

Certainty is calculated as the average of the digit probabilities of the
student ID number selected by the LAP solver. Thus, when LAP matches
Greedy, the certainties are the same, otherwise they differ.

The assemble_cost_matrix() is a helper function, so its invocation was
moved inside of lap_solver() in assign_prob.py, since lap_solver() needs
the heatmap probabilities as an argument for the certainty calculations.