Save QR image rotation angle in DB instead of on-disk modification
Works correctly, in the sense that if Web UI shows that an image needs a
90-degree CCW rotation to be upright, the image on-disk is sideways.

TODO: client is doing something wonky with rotations (perhaps not
accessing the DB field?). No guarantees that reassembly works correcly.
Furthermore, Web UI is strange, since the rotation of an image distorts
the div in which it is situated. Also need to test whether this works
correctly with the new image rotation command-line tool.