Removed extra field in dict breaking annotations
Previously I pulled this code out to make for a helper function. Then
with the addition of annotation-rubric links in the database I updated
the code to include an extra field in the dictionary. Currently it
is only used in this one case, so I reverted it back to the way it was
and removed the annotation portion that was breaking it.