Modify defaults for rubric kind, tags and meta fields
In legacy, the DB model specified that tags and meta fields default to ""

Tags field needs to default to "", because the annotator crashes if it
receives a None instead of a string (since it attempts to split the tags
field, believing it to be of type str).

Kind is a required field during rubric creation, so there isn't really a
point to setting a default empty string.