HW scribbler top box was too small vertically
Not sure why that started failing between pymupdf 1.23.4 and 1.23.5 but
it did look too small.  Increase height from 20 to 80 (it doesn't use
the vertical space if unneeded.

While at it, change the width to be based on the page size instead of a
number.

Leave some debugging code in place for drawing the rectangles, to assist
whoever next has to modify this!  And improve the error message to
describe what has failed.