Dudupe code, move PDF generator and fix timestamps
Remove one of the copies of the pdf gen code, and import from common
sources, by hacking some verbose and tqdm stuff.

The timestamps did not match the filename as generated seconds apart.
Change the the function to return a dict of information, including a
recommended filename and the timestamp.

Fix return type