Prevent unwinding bundle creation by racey signal handlers
Two line commit with a huge story...  TL-DR; is that the outer-most
atomic operation was rolling back under certain race conditions.  I
believe the signal handler (which is still susceptable to those race
conditions in the Huey Task ID setting) was unwinding the StagingBundle
creation, which the huey task then looked for.

I'm not even sure we really want that: maybe the topic level atomicity
is good: the entire bundle get's revoked... But for now, its inhibiting
debugging and replacing this race condition huey business.

I think this Fixes #2888.  That is the specific case of huey task rasising
StagingBundle.DoesNotExist.  Maybe also Fixes #3061.