Improve None handling
I played with the `pyright` type checker.  Here's a few fixes based on
what it complained about.  Some are legit: not checking for `None`
return on a function that very much could return `None`.

Others seem less legimate, like setting something inside a debug flag
and then later assuming---inside another `if`---that the something was
set.