Change ssl_verify to be a bool
We always used it as a bool.  But when you call `request` method of the
Requests library, it can be a string (path to a CA bundle), so I
documented that it could be a string.  However, other code (viz, in
chooser) seems to use it as a bool.

To decide what to do, I noticed that we use this value to set
`requestions.Session.verify` which is documented, at least in the
source, as being a bool.  So I think we can change it to bool in good
conscious.

Fixes #3279.