Raise PlomAPIException on 404 for server info
Getting the server info happens *before* we check the API version so the
usual error won't happen.  I think we originally thought we'd just
ignore that but we hit it today in the wild and wasted a bunch of
people's time.  So better to give a clearer error message about what
might be happening.

There was even a TODO in the code about the 404 case: so that has now
been addressed.

I think we cannot really tell the difference between 404 b/c its just
not a Plom server at all, and this error case, but that seems like a
less serious problem.  A timeout OTOH, will NOT be conflated: that is a
different error message.