Upload a toml file
I merged the separate toml-editor view htmx stuff: too hard for my brain
to understand even if it was nice to have more little files.

A remaining problem is that the text region does not update on file
upload: probably b/c htmx is used to update only the error messages.