pagescene: add a dryrun to delete for counting before we do it
I don't want to start a macro unless I have something to delete but I
want the delete logic in one place (as much as possible) so add a
dry-run.  We can iterate over adding things to a list then if that list
is non-empty, actually do some deletions inside a macro for Issue #3418.