small wait, ensure background procs still running
Very hard to be sure, b/c they could crash at any time, but if we wait
for a bit and they still poll without an exit condition, we can have
some confidence that its working.

This handles things like `python3 manageZZZZ.py ...` which fails
not quite immediately but very quickly.

Legacy used to try to ping the server in this case; and still it can be
racey/flaky.  IPC is hard.