server allow newer client, delegate responsibility
Leave it up to the client to decide if it can commnuicate with an older
server API, instead of forcing them to be identical.  Client Messenger
then keeps a minimum server API that it can support, which we currently
set to just its own local API version, but in the future we could allow
it to lag, and this would mean newer clients could connect to older
servers.

Fixes #3247.