scripts: do collectstatic whether or not we hotstart
This is b/c the hotstart is like possibly on a different container, and
we probably do not want to make the collected files persistent (b/c they
are versioned).  Fixes #3575.

This leaves a bug where the hot-started server will not have scrap and
extra pages: we'll solve that later.