Fix the ordering of FixedPagees
Fixes #3662.  Note there are two getters of such "page data": one of
them already sorted by page_number.  Ideally there would not be two:
that's a bigger refactor.

Document that the returned list is well-ordered (I had been assuming it
was arbitrary but various places assume the order means something and
present it to users (Marker -> View, Page Arrange in the absense of
other page_data, probably other places).  This also fixes #3662.

Note that the page data used for Annotator DOES NOT come from this
function (at least in most cases?) so I don't see how this bug (#3662)
could be causing #3659.  So this may not be a fix for 3659.