remove unused tracker and adjust error message
At least for read QR and bundle split, we do not explicitly track the
child chores.  I suppose this is b/c the parent chore is effectively (or
perhaps ineffectively---I'm still working through some bugs) tracking
them itself.  So we need to clarify an error message, noting that its ok
if the tracker does not exist.