Match filenames for downloading the qvmap
Command-line and web ui produce same filename by using a central
definition in the service.

Fixes #3726.

It was quite tricky to get it to download properly: needed to wrap the
text into bytes and then into a BytesIO.  Before this, the `filename=`
kwarg seemed to do nothing.  Also, the service was wrong: it was
returning a list of strings instead of a string despite its name.