Don't hardcode colour for sidebar links
The hardcoded black colour did not give enough contrast in the sidebar.
But without that custom CSS, the blue links where a bit overdone.  The
easiest fix for me was the make them all "secondary", which seems to
look ok in both light and dark colour schemes, at least using the
"manager" theme.