delete all calls delete one in a loop
It is true that this is slightly less efficient.  I don't care.  This is
tricky stuff, juggling the DB object delete and the file-store unlink:
so let's keep the correct code in just one place, even if theoretically
we make a couple extra DB lookups.  "No one will every have more than
640K versions".

I was worried about Issue #3825 when I did this fix, although it seems
that wasn't an actual issue.