Hide pip upgrade messages on old OS, min deps CI
In some cases, we'd like to know things work with older pip.  So we set
an env var to make `pip` not check versions.  Fixes #3877.

In contrast in jobs like `newOS_newdeps`, we'd like to know what breaks
on the latest pip.  In this case, we do upgrade `pip` as per Issue #3877.