Add key-value settings, use for fractional rubrics
To avoid changing the database a lot, let's just store things in a
key-value store backed by a json blob.  Probably slow if you hit it a
lot but I think it allows more nimble development.

Use this new key-value store to control whether fractional rubrics are
allowed or not.