Don't enforce integrity in HTML for self-hosted css/js
This seems like a bad thing but its a actually a good thing: we check
the hashes at download time.  Then we self-host these files.  So having
WET code means we might forget to update the HTML and then the rendering
would (presumably) break in production.  So we'd have to write some kind
of linter to make sure that doesn't happen...  but WE HOST THE FILES so
I don't see what we gain (except a foot gun) by having this integrity.

TL-DR; remove it.  Fixes #4031.