lead_marker not identifier by default, but promote does
"promote" action (button in UI) puts users into both "lead_marker" and
"identifier" group b/c that is what is used to do.  But in general not
all lead_markers are identifiers so remove it from the group
dependencies.  This allows you to remove the "identifier" abilility from
some "lead_markers".