allow possibility of knowing one of question, paper_num
The previous logic only checks paper_num b/c there was no way to set
paper_num but not question_idx_list.  It seems safer to express the full
logic, in case more general support is added in the future.  For
example, as user could tell us "this is in paper 17 but I haven't which
questions" or "This is definitely the answer to question 4 but I don't
know yet which paper number".  Both of those examples would give rise to
EXTRA pages that do not yet have (complete) data.